+++ src/app/app.component.ts
@@ -1,10 +1,33 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
+import * as fromRoot from '@app/root-store';
+import { select, Store } from '@ngrx/store';
+import { Observable } from 'rxjs';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })
-export class AppComponent {
-  title = 'ngrx-demo';
+export class AppComponent implements OnInit {
+  public title = 'ngrx-demo';
+  public config$: Observable<any[]>;
+  public errorMessage$: Observable<string>;
+  public isLoading$: Observable<boolean>;
+
+  constructor(private store$: Store<fromRoot.State>) { }
+
+  ngOnInit() {
+
+    this.isLoading$ = this.store$.pipe(
+      select(fromRoot.AppInitSelectors.selectIsLoading)
+    );
+
+    this.config$ = this.store$.pipe(
+      select(fromRoot.AppInitSelectors.selectConfig)
+    );
+
+    this.errorMessage$ = this.store$.pipe(
+      select(fromRoot.AppInitSelectors.selectError)
+    );
+  }
 }