+++ src/app/core/core.module.ts
@@ -2,14 +2,15 @@ import { CommonModule } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
 import { NgModule, Optional, SkipSelf } from '@angular/core';
 import { HeaderComponent } from '@core/header/header.component';
+import { LoaderComponent } from '@core/loader/loader.component';
 
 @NgModule({
   imports: [
     CommonModule,
     HttpClientModule
   ],
-  declarations: [HeaderComponent],
-  exports: [HeaderComponent],
+  declarations: [HeaderComponent, LoaderComponent],
+  exports: [HeaderComponent, LoaderComponent],
   providers: []
 })
 export class CoreModule {