+++ src/app/core/loader/loader.component.ts
@@ -1,0 +1,16 @@
+import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
+
+@Component({
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  selector: 'app-loader',
+  templateUrl: './loader.component.html',
+  styleUrls: ['./loader.component.css']
+})
+export class LoaderComponent {
+
+  @Input() public errorMessage: string;
+  @Input() public isLoading: boolean;
+
+  constructor() { }
+
+}