+++ src/app/root-store/root-store.effects.ts
@@ -1,13 +1,13 @@
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
 import * as rootStoreActions from '@app/root-store/root-store.actions';
-import { Actions, Effect, ofType } from '@ngrx/effects';
+import { Actions, Effect, ofType, OnInitEffects, ROOT_EFFECTS_INIT } from '@ngrx/effects';
 import { Action } from '@ngrx/store';
 import { Observable } from 'rxjs';
 import { map, tap } from 'rxjs/operators';
 
 @Injectable()
-export class RootStoreEffects {
+export class RootStoreEffects implements OnInitEffects {
 
   constructor(
     private actions$: Actions,
@@ -21,5 +21,23 @@ export class RootStoreEffects {
     tap(() => { this.router.navigate(['basic']); }),
     map(() => (new rootStoreActions.LogoutToClearStore()))
   );
+/*
+  ROOT_EFFECTS_INIT gets triggered before  APP_INITIALIZER complete its task.
+  Here is the work around
+  https://github.com/ngrx/platform/issues/931
+*/
+  @Effect({ dispatch: false })
+  init$ = this.actions$.pipe(
+    ofType(ROOT_EFFECTS_INIT),
+    tap(() => console.log('All the Root Effects has been added and registered'))
+  );
+
+  /*
+  OnInitEffects gets invoked before ROOT_EFFECTS_INIT.
+  */
+  ngrxOnInitEffects(): Action {
+    console.log('Root Effects has been Initialized');
+    return { type: '[RootEffects]: Init' };
+  }
 
 }